Fixes issue with Daylight Saving Time for "Floating" Due dates.
A Floating Due Date doesn't have a TimeZone. So the "default time zone offset" was being appended.

There were 2 problems with this:
1. The "default timezone offset" was computed only once on startup.
So the default timezone offset wouldn't change when Daylight Saving Time changed.

2. The other issue is that the correct timezone offset to use actually depends on the date of the due date.

So this change updates it to calculate the "default timezone offset" at the date of the due date.

With this, Floating Due Dates should now show up correctly.