Adds support for Todoist Durations
At some point after I created this program (which depended on certain Labels to specify the duration of an event), Todoist added Durations to Tasks.

So now, we will still use labels if they are defined, but if a Task has a Duration defined, it will override any duration labels.

This gives us a lot more flexibility in defining the duration of a Task. And it is also simpler to create tasks with durations in Todoist.